Refactor sleep
