+++ packages/coreutils/src/index.ts
@@ -15,3 +15,4 @@ export * from './signal';
 export * from './text';
 export * from './time';
 export * from './url';
+export * from './utils';
